refactor: extract provider_names module so low-level code doesn't import from FastAPI routes
Address Copilot review on PR #169: `cloud_storage.py` (and `cloud_client.py`)
were importing `to_wire_provider` / `from_wire_provider` from
`integration_routes` at runtime, creating an unnecessary dependency from
low-level utilities onto the FastAPI routes module. Move both helpers to
`provider_names.py` and update the three call sites.
